pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pymoney"
version = "0.1.0"
description = "Personal finance system"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=0.10",
    "marimo>=0.3",
    "plotly>=5.0",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pytest>=8.0",
    "pyyaml>=6.0",
    "gspread>=6.0",
    "google-auth>=2.0",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
pymoney = "pymoney.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
